+++ spec/middleware_spec.rb
@@ -2,6 +2,7 @@
 require 'spec_helper'
 require 'rails_multisite'
 require 'rack/test'
+require 'json'
 
 describe RailsMultisite::Middleware do
   include Rack::Test::Methods
@@ -20,6 +21,11 @@ describe RailsMultisite::Middleware do
         run (proc do |env|
           request = Rack::Request.new(env)
           [200, { 'Content-Type' => 'text/html' }, "<html><BODY><h1>#{request.hostname}</h1></BODY>\n \t</html>"]
+        end)
+      end
+      map '/salts' do
+        run (proc do |env|
+          [200, { 'Content-Type' => 'application/json' }, env.slice(*RailsMultisite::CookieSalt::COOKIE_SALT_KEYS).to_json]
         end)
       end
     }.to_app
@@ -96,6 +102,25 @@ describe RailsMultisite::Middleware do
     it 'returns 404 for invalid site' do
       get '/html'
       expect(last_response).to be_not_found
+    end
+  end
+
+  describe 'encrypted/signed cookie salts' do
+    it 'updates salts per-hostname' do
+      get 'http://default.localhost/salts'
+      expect(last_response).to be_ok
+      default_salts = JSON.parse(last_response.body)
+      expect(default_salts.keys).to contain_exactly(*RailsMultisite::CookieSalt::COOKIE_SALT_KEYS)
+      expect(default_salts.values).to all(include("default.localhost"))
+
+      get 'http://second.localhost/salts'
+      expect(last_response).to be_ok
+      second_salts = JSON.parse(last_response.body)
+      expect(second_salts.keys).to contain_exactly(*RailsMultisite::CookieSalt::COOKIE_SALT_KEYS)
+      expect(second_salts.values).to all(include("second.localhost"))
+
+      leaked_previous_hostname = second_salts.values.any? { |v| v.include?("default.localhost") }
+      expect(leaked_previous_hostname).to eq(false)
     end
   end
 end